+++ src/clientRoutes.ts
@@ -61,10 +61,6 @@ const handleCreateWalletWithKeychains = function(req) {
   return req.bitgo.wallets().createWalletWithKeychains(req.body);
 };
 
-const handleEthGenerateWallet = function(req) {
-  return req.bitgo.eth().wallets().generateWallet(req.body);
-};
-
 const handleSendCoins = function(req) {
   return req.bitgo.wallets().get({ id: req.params.id })
   .then(function(wallet) {
@@ -103,17 +99,6 @@ const handleCreateTransaction = function(req) {
   return req.bitgo.wallets().get({ id: req.params.id })
   .then(function(wallet) {
     return wallet.createTransaction(req.body);
-  })
-  .catch(function(err) {
-    err.status = 400;
-    throw err;
-  });
-};
-
-const handleEthSendTransaction = function(req) {
-  return req.bitgo.eth().wallets().get({ id: req.params.id })
-  .then(function(wallet) {
-    return wallet.sendTransaction(req.body);
   })
   .catch(function(err) {
     err.status = 400;
@@ -570,10 +555,6 @@ exports = module.exports = function(app, args) {
 
   app.put('/api/v1/wallet/:id/consolidateunspents', parseBody, prepareBitGo(args), promiseWrapper(handleConsolidateUnspents, args));
   app.put('/api/v1/wallet/:id/fanoutunspents', parseBody, prepareBitGo(args), promiseWrapper(handleFanOutUnspents, args));
-
-  // eth
-  app.post('/api/v1/eth/wallet/generate', parseBody, prepareBitGo(args), promiseWrapper(handleEthGenerateWallet, args));
-  app.post('/api/v1/eth/wallet/:id/sendtransaction', parseBody, prepareBitGo(args), promiseWrapper(handleEthSendTransaction, args));
 
   // any other API call
   app.use('/api/v[1]/*', parseBody, prepareBitGo(args), promiseWrapper(handleREST, args));
